Feat: Export functionality Refactor done
